Moved all the security report components into their own folder, finished folder structure
