+++ monkey/monkey_island/cc/ui/src/components/report-components/security/ScannedServers.js
